Rename module 'evaluation' to 'assessment'
